Add support for transfers on the local machine, this fixes #3
The README is updated to explain how to invoke local transfers on the current machine. TL;DR: Just omit the `hostname` key of an entry in the config file.

Also fixed a dict bug which made the program unable to run with Python3. Now, Python3 is supported fully. (Well… full… I'm using it with Python3 and haven't noticed any bugs.)