Add status output for monitoring mode
This is even given without --verbose. I think it has a different level than all the information printed with --verbose; but I also don't want to introduce verbosity levels. So, this is the compromise. An alternative could be an explicit --quiet mode, which would turn off all output. Let me know if you find the current behavior bad and/or would like to have --quiet.