Properly configur setuptools setup() to generate console script; add infrastructure
The old entry in setup.py, scripts=…, is still there, but is going to be deleted at one point. Right now a warning message is printed. To propagate Click's environment variables, the console script does not call main() directly, but through a wrapper; setuptools itself can not invoke main() with a variable…