Restructure and prepare for multi-hosts
Restructure: All functionality is (on the way to being sorted into) classes. There's a class holding global state (configParameters) and individual target state (syncObject) [names might change in the future]. There are some simplications and absorptions of global methods still to be done, but I'm half-way there.

Multi-Hosts: In the future, RBC will support synchronizing with multiple targets at once (hopefully…). Currently, multiple hosts are parsed successfully, but not handled any further (because the lambda function is not as easy).